# Faster load function and more omogeneus file parser interface