Hi, I found myself in the need to load thousands of pdb or gro files all in once and I noticed that the mdtraj.load function can be sped up without losing its high level functionalities

And I also noted that if using pdb or gro files (or in general files in the _TOPOLOGY_EXTS category) the topology gets re-parsed again for each single file.
Therefore I made the PDB and the GRO file parsers able to accept the top argument and therefore re-use the already parsed topology (I would suggest to slowly make the interface more omogeneus by implementing it in all the file parsers in _TOPOLOGY_EXTS, see #1649)

I made the load function able to deal more flexibly with the fact that some parsers may not accept top so if in the future someone implements the top keyword in other parsers it won't be necessary to change load

And changed the if order in _parse_topology in order for it to be faster if a Topology or a Trajecotry is given

The speed up is pretty good for lists of files (12% faster):
```
Before

python -m timeit -n 10 -s 'import mdtraj' -- 'mdtraj.load("tests/data/1am7_protein.pdb", top="tests/data/1am7_protein.pdb")'

10 loops, best of 5: 517 msec per loop



python -m timeit -n 2 -s 'import mdtraj; list=["tests/data/1am7_protein.pdb"]*10' -- 'mdtraj.load(list, top="tests/data/1am7_protein.pdb")'

2 loops, best of 5: 3.17 sec per loop


After

python -m timeit -n 10 -s 'import mdtraj' -- 'mdtraj.load("tests/data/1am7_protein.pdb", top="tests/data/1am7_protein.pdb")'

10 loops, best of 5: 460 msec per loop


python -m timeit -n 2 -s 'import mdtraj; list=["tests/data/1am7_protein.pdb"]*10' -- 'mdtraj.load(list, top="tests/data/1am7_protein.pdb")'

2 loops, best of 5: 2.63 sec per loop
```

ps if the PR is accepted could it be possible to include it in the 1.9.6 release or is it too late?