@MauriceKarrenbrock 2021-05-14
I have tried this new stuff in a real life situation and the speed up is actually much greater than 12%
One process that took me more than one hour is now taking me more or less 10 minutes