@MauriceKarrenbrock 2021-05-17
I am sorry I think I removed your comment about the deepcopy on the code as an error.

In any case what happens to you is normal, atom indexes are always zero indexed and select creates a new topology. When you where selecting atom 1 on a 2 atoms topology you were looking for a non existent second atom and therefore you got an empty output topology

In any case I don't know of the deepcopy is actually necessary in order not to modify the input topology, any opinions? (I am always unsure on how python deals with pointers to objects in functions)