@sroet 2021-05-17
> In any case what happens to you is normal, atom indexes are always zero indexed and select creates a new topology. When you were selecting atom 1 on a 1 atom topology you were looking for a non existent second atom and therefore you got an empty output topology

Yes, I understand that that is what is supposed to happen, I am just very worried that a user might do:
```python
# Say I am only interested even indices in a pdb with 8 atoms
frame0 = mdtraj.load("test1.pdb", atom_indices=[0,2,4,6])
# frame0.topology is now a subsetted topology 
# Now use the loading trick as you describe (don't build a new top)
frame1 = mdtraj.load("test2.pdb", atom_indices=[0,2,4,6], top=frame0.topology)
```
This leads would lead to `frame1` having a broken topology with only information for atom `0` and `6` (the even atoms in the subsetted topology) while the `xyz` contains the info for atoms `0`,`2`,`4`,`6`. Personally, I would expect that a given top would never be altered (even a copy inside an instance). 

> In any case I don't know of the deepcopy is actually necessary in order not to modify the input topology, any opinions? (I am always unsure on how python deals with pointers to objects in functions)

Python is mostly pass by reference, so I would keep the deepcopy to be safe