@MauriceKarrenbrock 2021-05-17
I tried to do exactly as you said and I got an exception
```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/media/maurice/DATI/Documenti/maurice/Phd_unige/mdtraj/mdtraj/core/trajectory.py", line 429, in load
    t = loader(tmp_file, **kwargs)
  File "/media/maurice/DATI/Documenti/maurice/Phd_unige/mdtraj/mdtraj/formats/pdb/pdbfile.py", line 189, in load_pdb
    unitcell_angles=unitcell_angles)
  File "/media/maurice/DATI/Documenti/maurice/Phd_unige/mdtraj/mdtraj/core/trajectory.py", line 1202, in __init__
    self.xyz = xyz
  File "/media/maurice/DATI/Documenti/maurice/Phd_unige/mdtraj/mdtraj/core/trajectory.py", line 916, in xyz
    warn_on_cast=False, add_newaxis_on_deficient_ndim=True)
  File "/media/maurice/DATI/Documenti/maurice/Phd_unige/mdtraj/mdtraj/utils/validation.py", line 148, in ensure_type
    raise error
ValueError: xyz must be shape (Any, 4, 3). You supplied  (1, 22, 3)
```

So I guess that we don't need to worry (I used native.pdb from the tests/data directory)