@sroet 2021-05-17
Alright, I had some time to properly run through your code.

There is some feature regression with this:
master:
```python
>>> import mdtraj as md
>>> frame0 = md.load(["native.pdb"], atom_indices=[0,2,4,6])
>>> frame1 = md.load(["native.pdb"], atom_indices=[0,2,4,6], top=frame0.topology)
/home/sander/miniconda3/envs/python3/lib/python3.7/site-packages/mdtraj/core/trajectory.py:422: UserWarning: top= kwarg ignored since file contains topology information
  warnings.warn('top= kwarg ignored since file contains topology information')
```
this PR:
```python
>>> import mdtraj as md
>>> md.load(["native.pdb", "native.pdb"], atom_indices=[0,2,4,6])
<mdtraj.Trajectory with 2 frames, 4 atoms, 2 residues, without unitcells at 0x7eff8e6c59d0>
>>> frame0 = md.load("native.pdb", atom_indices=[0,2,4,6])
>>> frame0
<mdtraj.Trajectory with 1 frames, 4 atoms, 2 residues, without unitcells at 0x7eff8e6aefa0>
>>> frame1 = md.load("native.pdb", atom_indices=[0,2,4,6], top=frame0.topology)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/sander/github_files/mdtraj/mdtraj/core/trajectory.py", line 429, in load
    t = loader(tmp_file, **kwargs)
  File "/home/sander/github_files/mdtraj/mdtraj/formats/pdb/pdbfile.py", line 187, in load_pdb
    traj = Trajectory(xyz=coords, time=time, topology=topology,
  File "/home/sander/github_files/mdtraj/mdtraj/core/trajectory.py", line 1202, in __init__
    self.xyz = xyz
  File "/home/sander/github_files/mdtraj/mdtraj/core/trajectory.py", line 915, in xyz
    value = ensure_type(value, np.float32, 3, 'xyz', shape=shape,
  File "/home/sander/github_files/mdtraj/mdtraj/utils/validation.py", line 148, in ensure_type
    raise error
ValueError: xyz must be shape (Any, 2, 3). You supplied  (1, 4, 3)
```
Which might or might not be acceptable. 
In my opinion the error now is very strange/wrong, as I gave a topology with 4 atoms, but now the xyz only expects 2.

However, this might be solvable with proper documentation warning the user that the topology has to be correct for the whole loaded pdb, not just the atom_indices

@MauriceKarrenbrock  Another speedup that you might want to look into is to prevent `subset` or `copy` of topologies. [In a project of mine](https://github.com/dwhswenson/contact_map/pull/35#issuecomment-389322129), I discovered that subsetting is reasonably slow (in the order of 0.4 s for a 2700 atom system), and it scales both with the number of atoms in the full topology and the number of atoms after slicing. 

If you want to prevent this overhead and you feel that this is in the scope of this PR, then you can add:

```python
    # Little monkey-patches to prevent further subsetting Topologies                    
    top = subset_topology.copy()                                                
    # Override copy to just return self after this  (this is called by deepcopy)                                  
    top.copy = type(top.copy)(lambda self: self, top)                           
    # Override subset to just return self                                             
    top.subset = type(top.subset)(lambda self, atom_indices: self, top)         
    kwargs['top'] = top                                                             
```
 before the `for f in filename_or_filenames` loop in `trajectory.py`