@MauriceKarrenbrock 2021-05-17
I made the exception more informative in case one introduces a mismatching topology so there should be less problems of clarity and added a warning when the file parser doesn't support top in the kwargs

For the second thing I see the advantages but as the final trajectory will contain this modified topology there will be some hard to catch bugs on the user side (the user thinks he is copying or subsetting but in reality he is not) therefore I wouldn't insert it in the load function, but it might be added as a suggestion in the documentation under `Notes` so that a user can overwrite the subset method of the input topology if needed