@sroet 2021-05-17
> For the second thing I see the advantages but as the final trajectory will contain this modified topology there will be some hard to catch bugs on the user side (the user thinks he is copying or subsetting but in reality he is not) therefore I wouldn't insert it in the load function, but it might be added as a suggestion in the documentation under Notes so that a user can overwrite the subset method of the input topology if needed.

Ah, but that broken topology is always overwritten with the correct `subset_topology` inside your load-loop, so it should never exit that function